Add version url config workaround
This subclasses KSC's generic Password plugin to allow version discovery with
default Keystone configurations that leave admin_endpoint and public_endpoint
at the default values (http://localhost:xxxx).  This patch copies the scheme
and netloc from the original auth_url into the URL returned from version
discovery if the returned netloc begins with 'localhost'.

Due to the specific nature of this review, the Keystone team is not
inclned to include it in keystoneclient so it is addressed here.

Closes-bug: #1410364
Change-Id: I877fe74d86aab3a63122a07b77d1302a007f5b30