temporarily skip help tests
We're trying to change cliff for the better, by improving the UX
of the help command. But cliff tests against the tip of OSC.

Temporarily skip the tests that are failing and then update them
once the fix for cliff has been merged.

The patch in question: https://review.openstack.org/#/c/201258/

Change-Id: Ie1e7ddd66477ddf4a3bafedad12934e900ace48f