Attempt to find resource by ID, without kwargs
In keystone, to look up resources we just need the ID. Supplying
query arguments in a get() request to the client will result in
a list being performed. The resulting query will always come up
with no hits since it'll be trying to find the user "name" even
if an ID is given.

Change-Id: I111b2730eefec83968ee89aa4a56cb518ca9b203
Closes-Bug: #1597246