Avoid string actions on non-string objects
When config is coming from a clouds.yaml file, Python's yaml parser may
turn the version into an int if it is not quoted in the yaml. Ensure
that we set the value to a string before interacting with it.

Also fix a possible string action on a None object.

Change-Id: Ibf02a85d89c95ba76688b3f5c417f861173008d5
Closes-Bug: #1634986