Add Constraints support
Adding constraints support to libraries is slightly more complex than
services as the libraries themselves are listed in upper-constraints.txt
which leads to errors that you can't install a specific version and a
constrained version.

This change adds constraints support by also adding a helper script to
edit the constraints to remove osc-lib.

Change-Id: Ie6b8f7a7c46becc80d6cf9251fbd5eb5c9c707b1