Revert "Update UPPER_CONSTRAINTS_FILE for feature/osc4"
There is no feature/osc4 branch in openstack/requirements repository.
osc-lib itself has no dependency on a specific branch related to
feature/osc4, so this is no need for this kind of stuffs.

This change breaks tox jobs when running them in local env.

Note that the upstream gate succeeded because UPPER_CONSTRAINTS_FILE
envvar is specified in our gate job.

This reverts commit ef967cbeba4ea452c9a145a6318b4120a48433d7.

Change-Id: I3acacbb1d4190a18f05838f1aea4c37254b50b10