Add project lookup utils
It is not special to specify a project of a resource.
At now, find_project() is defined in OSC, but OSC plugins
would like to consume it.

The original idea is based on openstackclient.identity.common,
but the proposed version is re-implemented on top of OpenStack SDK.
sdk_connection is saved in the clientmanager in the new stype of
OpenStack SDK connection and the commit assumes it.
The feature depends on the OpenStack SDK patch which allows us
to pass domain information to find_project().

Also adds add_project_owner_option_to_parser() to add CLI options
(--project and --project-domain). It will help us make these options
consistent across OSC and OSC plugins.

Closes-Bug: #1632147
Depends-On: I60a8b3b83f6170b60d09c101b5c7035148283678
Change-Id: I8f59fa3f9b7c573485cd1572e5e9aae08f071e37