Relocate utils tests to match the main code
commit 75700a7b4247773cf1f919b40e70ba89ba92bddb introduced
utils/ directory in the main code, but the test code was not relocated.
Locations of test codes should match corresponding main codes.
This commit relocates utils test codes.

Note that osc_lib.tests.utils is a module exposed to osc-lib consumers,
so it is moved to osc_lib/tests/utils/__init__.py not to break
existing imports.

Change-Id: Ib134ba6568396dcb9c6502c3f8f292823fdc1564