Suppress subTest() expected errors
Prior to this patch, calling subTest from a test inside a class
that inherits from osc_lib.TestCase would have failed the test.
This failure is either due to lack of support prior to PY3.4 [1], or
long-standing bug in testtools repo [2] which TestCase inherits from.

The patch suppresses both errors, so to allow writing better tests
using subTest, despite the fact that it is currently broken.

[1] - https://docs.python.org/3/whatsnew/3.4.html#unittest
[2] - https://bugs.launchpad.net/testtools/+bug/1219404

Partial-Bug: #1742343

Change-Id: I56bb3501d7a33fcec4fd971b2d9685d113e0af43