Add cloud to warning about domain related config
When this warning fires you get a bunch of

 Ignoring domain related config project_domain_name because identity API version is 2.0
 Ignoring domain related config user_domain_name because identity API version is 2.0

in the output, but no clue as to what cloud was actually causing the
problem to help you debug.  This is a bit annoying if you have many
different clouds setup in your configuration.  Add the name of the
cloud to the message.

Change-Id: Id2b5133f583f7071745e02001eaeaa3dcee211ba