Fix exception in log warning when no cloud config exists
A recent patch added a more descriptive log message, which is great.
Unfortunately, config is not guaranteed in this case to have a cloud
key, so just accessing it with config['cloud'] throws an exception,
which in turn breaks a test in python-openstackclient.

Expand the code a little bit to log the warning differently depending on
whether or not cloud exists and has a value.

Change-Id: I6b058b3c48a6018b1fe8084344f370530d730d25