switch documentation job to new PTI
This is a mechanically generated patch to switch the documentation
jobs to use the new PTI versions of the jobs as part of the
python3-first goal.

See the python3-first goal document for details:
https://governance.openstack.org/tc/goals/stein/python3-first.html

Update docs tox environment to build ChangeLog file when using tox -e
docs.

Change-Id: If68bdb33bbf0727d699268d9de511790c8a048be
Story: #2002586
Task: #24320