Mask password when specified as an argument with a space
I was using a command line like

 $ openstack user password set --password "p" --original-password="x"

and saw the --password argument echoed in the info log

 START with options: ['-v', ..., 'user', 'password', 'set', '--password', 'p', '--original-password=***']

The masking works if you set "--password=p", but not "--password p".
This is because the password matching regex doesn't know they're
related.  An easy way to avoid this is to just make the whole log line
back into a single string, and this gets masked correctly.

Change-Id: I0fdd895c9449d9a90d811af4ded34f32ba2b2215