Default to using public interface for all clients
Within osc_lib, we currently set the default interface (for
OS_INTERFACE/--os-interface) to None.  This results in using the
defaults from each client.  In nearly all clients, the default is
to use the 'public' interface.  In keystoneclient, the default is
'admin'.  This means that end users using a CLI like 'openstack'
will need to have access to the admin endpoint to perform any
operation against the identity API unless they specify --os-interface
to explicitly use the public interface themselves.  This does not
provide a nice usability experience.

This patch changes the default interface to 'public' to enforce a
consistent experience across all clients with regards to the interface
that is used.  The default can still be overridden via the existing
--os-interface/OS_INTERFACE methods.

Change-Id: I4562a6a4bf5318d05ed9ccc347694d730125e4cc