Move the _tag module from OSC to osc-lib
The neutron code in the OpenStack Client(OSC) includes a module
for handling tags[1]. This module is useful for other projects that
support API tags like neutron does.

This is a lift and shift of the _tag.py from the neutron code
in OSC to the osc-lib so that other projects can use it.
In the process I added doc strings and unit test coverage.

A follow on patch will switch the neutron code to use the
module from osc-lib.

[1] https://opendev.org/openstack/python-openstackclient/src/ \
    branch/master/openstackclient/network/v2/_tag.py

Change-Id: Icdfdcf3a3ed657d92766081e138733ff43f47f38