format_dict() returns no value for None
We have some subtle cases to distinguish:
* value of None   (the Python literal)
* value of ''     (an empty string)
* value of 'None' (a string)

Return no representation of data when the value is the Python literal None
rather than the string 'None' which can be confused for an actual string value.

Returning an empty string ('') can be confused for an actual empty
string so that is also not an option.  By leaving off the quotes we can accurately
reflect the distinct value of None.

Change-Id: Icdc1712e626a7e8a706348586a6be2e7612c4376
Signed-off-by: Dean Troyer <dtroyer@gmail.com>