Save the results from the early authentication
Somehow we lost the benefit of doing early authentication forcing
some of the project clients (particularly Identity) to do it again.
Saving the auth_ref in ClientManager is not enough, we need to stuff
it back into the Session too so it is also available to the non-SDK
clients.

Arguably this could be fixed in ClientManager.auth_ref itself but I
am not certain that is a desirable side-effect.  This solution
addresses the immediate concern.

Change-Id: Ie7aca398bba6fba30ef1b432daaefcd7d8b780e0
Signed-off-by: Dean Troyer <dtroyer@gmail.com>