formatters: Cast *all* columns before outputting
This is a continuation of a previous change, change
Ib7db6a25f2352a013cb2ce603e60ca48b6cc70e6 ("formatters: Cast columns
before outputting"). As noted in that change, libraries can return
non-primitive types for many operations which can break YAML output
formatting since PyYAML only works with primitive types in safe mode.
That change was not complete as it only handled the formatters for dicts
and lists, not dicts of lists or lists of dicts. Close this gap now.

While we're here, we improve test coverage and catch a number of bugs
with the existing tests.

Change-Id: Ifd8db5b725f46a1ca216986167a0bcde5aec1e16
Signed-off-by: Stephen Finucane <stephenfin@redhat.com>